클라우드 구성을 위한 내부 노드 단일화 및 소스 정리
