+++ src/assets/scss/_designer-theme-default.scss
@@ -1,7 +1,7 @@
 :root {
   --vf-node-bg: #fff;
   --vf-node-text: #222;
-  --vf-connection-path:  #b1b1b7;
+  --vf-connection-path: #b1b1b7;
   --vf-handle: #555;
 }
 
@@ -58,13 +58,8 @@
 .vue-flow__node-input,
 .vue-flow__node-output,
 .vue-flow__node-cloud,
-.vue-flow__node-master,
-.vue-flow__node-worker,
-.vue-flow__node-loadbalancer,
-.vue-flow__node-registry,
-.vue-flow__node-storageserver,
-.vue-flow__node-storagecluster,
-.vue-flow__node-etcdcluster {
+.vue-flow__node-capi,
+.vue-flow__node-member {
   padding: 10px;
   border-radius: 3px;
   // width: 150px;
@@ -79,9 +74,9 @@
 
   &.selected,
   &.selected:hover {
-     box-shadow: 0 0 0 0.5px var(--vf-box-shadow);
-     border: 0.15rem solid #933
-   }
+    box-shadow: 0 0 0 0.5px var(--vf-box-shadow);
+    border: 0.15rem solid #933
+  }
 
   .vue-flow__handle {
     background: var(--vf-handle);
@@ -92,13 +87,8 @@
 .vue-flow__node-input.selectable,
 .vue-flow__node-output.selectable,
 .vue-flow__node-cloud.selectable,
-.vue-flow__node-master.selectable,
-.vue-flow__node-worker.selectable,
-.vue-flow__node-loadbalancer.selectable,
-.vue-flow__node-registry.selectable,
-.vue-flow__node-storageserver.selectable,
-.vue-flow__node-storagecluster.selectable,
-.vue-flow__node-etcdcluster.selectable {
+.vue-flow__node-capi.selectable,
+.vue-flow__node-member.selectable {
   &:hover {
     box-shadow: 0 1px 4px 1px rgba(0, 0, 0, 0.08);
   }
@@ -129,13 +119,8 @@
 }
 
 .vue-flow__node-cloud,
-.vue-flow__node-master,
-.vue-flow__node-worker,
-.vue-flow__node-loadbalancer,
-.vue-flow__node-registry,
-.vue-flow__node-storageserver,
-.vue-flow__node-storagecluster,
-.vue-flow__node-etcdcluster {
+.vue-flow__node-capi,
+.vue-flow__node-member {
   --vf-node-color: #0041d0;
   --vf-handle: var(--vf-node-color);
   --vf-box-shadow: var(--vf-node-color);
@@ -194,225 +179,3 @@
     }
   }
 }
-
-
-// :root {
-//   --vf-node-bg: #fff;
-//   --vf-node-text: #222;
-//   --vf-connection-path: #b1b1b7;
-//   --vf-handle: #555;
-//   --vf-handle-source: gray;
-//   --vf-handle-target: #F00;
-// }
-
-// .vue-flow__selection {
-//   background: rgba(0, 89, 220, 0.08);
-//   border: 1px dotted rgba(0, 89, 220, 0.8);
-// }
-
-// .vue-flow__edge {
-//   &.selected {
-//     .vue-flow__edge-path {
-//       stroke: #933;
-//     }
-//   }
-
-//   &.animated path {
-//     stroke-dasharray: 5;
-//     animation: dashdraw 0.5s linear infinite;
-//   }
-
-//   &.updating {
-//     .vue-flow__edge-path {
-//       stroke: #777;
-//     }
-//   }
-// }
-
-// .vue-flow__edge-path {
-//   stroke: #b1b1b7;
-//   stroke-width: 2;
-// }
-
-// .vue-flow__edge-text {
-//   font-size: 10px;
-// }
-
-// .vue-flow__edge-textbg {
-//   fill: #fff;
-// }
-
-// .vue-flow__connection-path {
-//   stroke: var(--vf-connection-path);
-//   stroke-width: 2;
-// }
-
-// .vue-flow__node {
-//   cursor: grab;
-//   resize: both;
-// }
-
-// .vue-flow__node-default,
-// .vue-flow__node-input,
-// .vue-flow__node-output,
-// .vue-flow__node-group,
-// .vue-flow__node-master,
-// .vue-flow__node-worker,
-// .vue-flow__node-cloud,
-// .vue-flow__node-cluster,
-// .vue-flow__node-router,
-// .vue-flow__node-lb {
-//   padding: 10px;
-//   border-radius: 3px;
-//   //width: 150px;
-//   font-size: 12px;
-//   color: var(--vf-node-text);
-//   text-align: center;
-//   border-width: 1px;
-//   border-style: solid;
-
-//   background: var(--vf-node-bg);
-//   border-color: var(--vf-node-color);
-
-//   &.selected,
-//   &.selected:hover {
-//     box-shadow: 0 0 0 0.5px var(--vf-box-shadow);
-//     border: 0.15rem solid  #933
-//   }
-
-//   .vue-flow__handle {
-//     background: var(--vf-handle);
-//   }
-
-//   .vue-flow__handle.source {
-//     background: var(--vf-handle-source);
-//   }
-//   .vue-flow__handle.target {
-//     background: var(--vf-handle-target);
-//   }
-
-// }
-
-// .vue-flow__node-default.selectable,
-// .vue-flow__node-input.selectable,
-// .vue-flow__node-output.selectable,
-// .vue-flow__node-group.selectable,
-// .vue-flow__node-master.selectable,
-// .vue-flow__node-worker.selectable,
-// .vue-flow__node-cloud.selectable,
-// .vue-flow__node-cluster.selectable,
-// .vue-flow__node-router.selectable,
-// .vue-flow__node-lb.selectable {
-//   &:hover {
-//     box-shadow: 0 1px 4px 1px rgba(0, 0, 0, 0.08);
-//   }
-// }
-
-// .vue-flow__node-input {
-//   --vf-node-color: #0041d0;
-//   --vf-handle: var(--vf-node-color);
-//   --vf-box-shadow: var(--vf-node-color);
-
-//   background: var(--vf-node-bg);
-//   border-color: var(--vf-node-color);
-// }
-
-// .vue-flow__node-default {
-//   --vf-node-color: #1a192b;
-//   --vf-handle: var(--vf-node-color);
-//   --vf-box-shadow: var(--vf-node-color);
-
-//   background: var(--vf-node-bg);
-//   border-color: var(--vf-node-color);
-// }
-// .vue-flow__node-master {
-//   --vf-node-color: #1a192b;
-//   --vf-handle: var(--vf-node-color);
-//   --vf-box-shadow: var(--vf-node-color);
-
-//   background: var(--vf-node-bg);
-//   border-color: var(--vf-node-color);
-// }
-// .vue-flow__node-worker {
-//   --vf-node-color: #1a192b;
-//   --vf-handle: var(--vf-node-color);
-//   --vf-box-shadow: var(--vf-node-color);
-
-//   background: var(--vf-node-bg);
-//   border-color: var(--vf-node-color);
-// }
-
-// .vue-flow__node-output {
-//   --vf-node-color: #ff0072;
-//   --vf-handle: var(--vf-node-color);
-//   --vf-box-shadow: var(--vf-node-color);
-// }
-
-// .vue-flow__node-group {
-//   background: lightgreen;
-//   --vf-node-color: green;
-//   --vf-handle: var(--vf-node-color);
-//   --vf-box-shadow: var(--vf-node-color);
-// }
-
-// .vue-flow__node-cloud {
-//   --vf-node-color: gray;
-//   --vf-handle: var(--vf-node-color);
-//   --vf-box-shadow: var(--vf-node-color);
-//   resize:both;
-// }
-
-// .vue-flow__node-cluster {
-//   --vf-node-color: gray;
-//   --vf-handle: var(--vf-node-color);
-//   --vf-box-shadow: var(--vf-node-color);
-// }
-
-// .vue-flow__nodesselection-rect {
-//   background: rgba(0, 89, 220, 0.08);
-//   border: 1px dotted rgba(0, 89, 220, 0.8);
-// }
-
-// .vue-flow__handle {
-//   position: absolute;
-//   width: 8px;
-//   height: 8px;
-//   background: var(--vf-handle);
-//   border: 1px solid #fff;
-//   border-radius: 100%;
-
-//   &.connectable {
-//     cursor: crosshair;
-//   }
-// }
-
-// .vue-flow__minimap {
-//   background-color: #fff;
-// }
-
-// .vue-flow__controls {
-//   box-shadow: 0 0 2px 1px rgba(0, 0, 0, 0.08);
-
-//   &-button {
-//     background: #fefefe;
-//     border-bottom: 1px solid #eee;
-//     box-sizing: content-box;
-//     display: flex;
-//     justify-content: center;
-//     align-items: center;
-//     width: 16px;
-//     height: 16px;
-//     cursor: pointer;
-//     user-select: none;
-//     padding: 5px;
-
-//     svg {
-//       max-width: 12px;
-//       max-height: 12px;
-//     }
-
-//     &:hover {
-//       background: #f4f4f4;
-//     }
-//   }
-// }
