+++ src/packages/designer/components/Nodes/Designer/MachineSetNode.vue
@@ -1,0 +1,9 @@
+<template>
+    <div></div>
+</template>
+
+<script setup lang="ts">
+</script>
+
+<style scoped lang="scss">
+</style>
