+++ src/packages/designer/composables/index.ts
@@ -4,4 +4,4 @@ export { default as useZoomPanHelper } from './useZoomPanHelper'
 export { default as useWindow } from './useWindow'
 export { default as useVueFlow } from './useVueFlow'
 // MOD: 클라우드 관리용
-export { default as useCloudHelper } from './useCloudHelper'
+export * from './useDesignerHelper'
