+++ src/packages/designer/types/index.ts
@@ -10,4 +10,4 @@ export * from './changes'
 export * from './handle'
 
 // MOD: Cloud 관리용
-export * from './cloud'
+export * from './designer'
